refactor(app): make main a thin entrypoint
