+++ src/lib.rs
@@ -1,0 +1,13 @@
+mod application;
+mod audio;
+mod core;
+mod input;
+mod local;
+mod platform;
+mod postprocess;
+mod runtime_config;
+mod session;
+mod text;
+mod transcriber;
+
+pub use application::run;
